% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdobject.R
\name{dump}
\alias{dump}
\title{Write updated R markdown file to disk}
\usage{
dump(rmdob, file)
}
\arguments{
\item{rmdob}{rmdmanip object}

\item{file}{Output file name}
}
\description{
Write updated R markdown file to disk
}
