% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdobject.R
\name{put}
\alias{put}
\title{Set root-level variable}
\usage{
put(rmdob, entry_name, entry_value)
}
\arguments{
\item{rmdob}{rmdmanip object}

\item{entry_name}{preamble yaml entry key}

\item{entry_value}{preamble yaml entry value}
}
\value{
rmdmanip object
}
\description{
Set root-level variable
}
