% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdobject.R
\name{put_param}
\alias{put_param}
\title{Set value in params block of R markdown preamble}
\usage{
put_param(rmdob, param_name, param_value)
}
\arguments{
\item{rmdob}{rmdmanip object}

\item{param_name}{key of the parameter}

\item{param_value}{value of the parameter}
}
\description{
Set value in params block of R markdown preamble
}
