% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdobject.R
\name{read.rmds}
\alias{read.rmds}
\title{Imports R markdown file from loaded string}
\usage{
read.rmds(string)
}
\arguments{
\item{string}{String representation of R markdown file}
}
\value{
rmdmanip object
}
\description{
Imports R markdown file from loaded string
}
